+++ packages/outbox/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.9.0
+
+### Minor Changes
+
+- refacto ServiceType --> STREAMS
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/database@3.1.0
+  - @volontariapp/shared@0.5.0
+  - @volontariapp/messaging@2.2.1
+
 ## 0.8.3
 
 ### Patch Changes